GH-2 Moved from vscode to Visual Studio
