--- MoneyCatApi/.vscode/tasks.json
@@ -1,15 +1,0 @@
-{
-    "version": "2.0.0",
-    "tasks": [
-        {
-            "label": "build",
-            "command": "dotnet",
-            "type": "process",
-            "args": [
-                "build",
-                "${workspaceFolder}/MoneyCatApi.csproj"
-            ],
-            "problemMatcher": "$msCompile"
-        }
-    ]
-}
