+++ Source/MoneyCat.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.421
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MoneyCat.API", "MoneyCat.API\MoneyCat.API.csproj", "{F3B592BF-5FDA-4B01-8FA0-5E1B5B775EB2}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F3B592BF-5FDA-4B01-8FA0-5E1B5B775EB2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F3B592BF-5FDA-4B01-8FA0-5E1B5B775EB2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F3B592BF-5FDA-4B01-8FA0-5E1B5B775EB2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F3B592BF-5FDA-4B01-8FA0-5E1B5B775EB2}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0B6ABB28-A900-4127-B6ED-D65303B28D61}
+	EndGlobalSection
+EndGlobal
